+++ cmd/vega/node/node_pre.go
@@ -68,10 +68,6 @@ func (l *NodeCommand) persistentPre(args []string) (err error) {
 
 	conf := l.confWatcher.Get()
 
-	if flagProvided("--no-chain") {
-		conf.Blockchain.ChainProvider = "noop"
-	}
-
 	// reload logger with the setup from configuration
 	l.Log = logging.NewLoggerFromConfig(conf.Logging)
 
@@ -104,9 +100,11 @@ func (l *NodeCommand) persistentPre(args []string) (err error) {
 
 	l.stats = stats.New(l.Log, l.conf.Stats, l.Version, l.VersionHash)
 
-	l.ethClient, err = ethclient.Dial(l.ctx, l.conf.NodeWallet.ETH.Address)
-	if err != nil {
-		return fmt.Errorf("could not instantiate ethereum client: %w", err)
+	if conf.Blockchain.ChainProvider != blockchain.ProviderNullChain {
+		l.ethClient, err = ethclient.Dial(l.ctx, l.conf.NodeWallet.ETH.Address)
+		if err != nil {
+			return fmt.Errorf("could not instantiate ethereum client: %w", err)
+		}
 	}
 
 	l.nodeWallets, err = nodewallets.GetNodeWallets(l.conf.NodeWallet, l.vegaPaths, l.nodeWalletPassphrase)
@@ -267,12 +265,13 @@ func (l *NodeCommand) startBlockchain(ctx context.Context, commander *nodewallet
 		l.checkpoint,
 		l.spam,
 		l.stakingAccounts,
+		l.rewards,
 		l.snapshot,
 		l.Version,
 	)
 
 	switch l.conf.Blockchain.ChainProvider {
-	case "tendermint":
+	case blockchain.ProviderTendermint:
 		srv, err := l.startABCI(ctx, app)
 		l.blockchainServer = blockchain.NewServer(srv)
 		if err != nil {
@@ -284,7 +283,7 @@ func (l *NodeCommand) startBlockchain(ctx context.Context, commander *nodewallet
 			return nil, err
 		}
 		l.blockchainClient = blockchain.NewClient(a)
-	case "nullchain":
+	case blockchain.ProviderNullChain:
 		abciApp := app.Abci()
 		n := nullchain.NewClient(l.Log, l.conf.Blockchain.Null, abciApp)
 
@@ -353,19 +352,29 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 	l.stakingAccounts, l.stakeVerifier = staking.New(
 		l.Log, l.conf.Staking, l.broker, l.timeService, l.witness, l.ethClient, l.netParams,
 	)
-
-	l.governance = governance.NewEngine(l.Log, l.conf.Governance, l.stakingAccounts, l.broker, l.assets, l.witness, l.netParams, now)
-
 	l.epochService = epochtime.NewService(l.Log, l.conf.Epoch, l.timeService, l.broker)
-	l.delegation = delegation.New(l.Log, delegation.NewDefaultConfig(), l.broker, l.topology, l.stakingAccounts, l.epochService, l.timeService)
+
+	if l.conf.Blockchain.ChainProvider == blockchain.ProviderNullChain {
+		// Use staking-loop to pretend a dummy builtin asssets deposited with the faucet was staked
+		stakingLoop := nullchain.NewStakingLoop(l.collateral, l.assets)
+		l.governance = governance.NewEngine(l.Log, l.conf.Governance, stakingLoop, l.broker, l.assets, l.witness, l.netParams, now)
+		l.delegation = delegation.New(l.Log, delegation.NewDefaultConfig(), l.broker, l.topology, stakingLoop, l.epochService, l.timeService)
+	} else {
+		l.governance = governance.NewEngine(l.Log, l.conf.Governance, l.stakingAccounts, l.broker, l.assets, l.witness, l.netParams, now)
+		l.delegation = delegation.New(l.Log, delegation.NewDefaultConfig(), l.broker, l.topology, l.stakingAccounts, l.epochService, l.timeService)
+	}
+
 	l.netParams.Watch(
 		netparams.WatchParam{
 			Param:   netparams.DelegationMinAmount,
 			Watcher: l.delegation.OnMinAmountChanged,
 		})
 
+	// setup rewards engine
+	l.rewards = rewards.New(l.Log, l.conf.Rewards, l.broker, l.delegation, l.epochService, l.collateral, l.timeService)
+
 	// checkpoint engine
-	l.checkpoint, err = checkpoint.New(l.Log, l.conf.Checkpoint, l.assets, l.collateral, l.governance, l.netParams, l.delegation, l.epochService)
+	l.checkpoint, err = checkpoint.New(l.Log, l.conf.Checkpoint, l.assets, l.collateral, l.governance, l.netParams, l.delegation, l.epochService, l.rewards)
 	if err != nil {
 		panic(err)
 	}
@@ -395,13 +404,10 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 		panic(err)
 	}
 
-	// setup rewards engine
-	l.rewards = rewards.New(l.Log, l.conf.Rewards, l.broker, l.delegation, l.epochService, l.collateral, l.timeService)
-
 	// notify delegation, rewards, and accounting on changes in the validator pub key
-	l.topology.NotifyOnKeyChange(l.delegation.ValidatorKeyChanged, l.stakingAccounts.ValidatorKeyChanged, l.rewards.ValidatorKeyChanged)
+	l.topology.NotifyOnKeyChange(l.delegation.ValidatorKeyChanged, l.stakingAccounts.ValidatorKeyChanged, l.rewards.ValidatorKeyChanged, l.governance.ValidatorKeyChanged)
 
-	l.snapshot.AddProviders(l.checkpoint, l.collateral, l.governance, l.delegation, l.netParams, l.epochService, l.assets, l.banking,
+  	l.snapshot.AddProviders(l.checkpoint, l.collateral, l.governance, l.delegation, l.netParams, l.epochService, l.assets, l.banking,
 		l.notary, l.spam, l.rewards, l.stakingAccounts, l.stakeVerifier, l.limits, l.topology, l.evtfwd, l.executionEngine)
 
 	// now instantiate the blockchain layer