+++ processor/abci.go
@@ -3,6 +3,7 @@ package processor
 import (
 	"bytes"
 	"context"
+	"encoding/base64"
 	"encoding/hex"
 	"errors"
 	"fmt"
@@ -106,6 +107,7 @@ type App struct {
 	netp            NetworkParameters
 	oracles         *Oracle
 	delegation      DelegationEngine
+	rewards         RewardEngine
 	limits          Limits
 	stake           StakeVerifier
 	stakingAccounts StakingAccounts
@@ -144,6 +146,7 @@ func NewApp(
 	checkpoint Checkpoint,
 	spam SpamEngine,
 	stakingAccounts StakingAccounts,
+	rewards RewardEngine,
 	snapshot Snapshot,
 	version string, // we need the version for snapshot reload
 ) *App {
@@ -184,6 +187,7 @@ func NewApp(
 		spam:            spam,
 		stakingAccounts: stakingAccounts,
 		epoch:           epoch,
+		rewards:         rewards,
 		snapshot:        snapshot,
 		version:         version,
 	}
@@ -456,7 +460,27 @@ func (app *App) OnEndBlock(req tmtypes.RequestEndBlock) (ctx context.Context, re
 	if app.spam != nil {
 		app.spam.EndOfBlock(uint64(req.Height))
 	}
-	return
+
+	// check if we need to update the voting power of the validators
+	if validatorUpdates := app.rewards.EndOfBlock(req.Height); validatorUpdates != nil {
+		vUpdates := make([]tmtypes.ValidatorUpdate, 0, len(validatorUpdates))
+		for _, v := range validatorUpdates {
+			// using the default for key type which is ed25519 which seems fine for now because all validators in genesis use this key type
+			// TODO use the proper key type of the validators
+			pubkey, err := base64.StdEncoding.DecodeString(v.TmPubKey)
+			if err != nil {
+				continue
+			}
+			update := tmtypes.UpdateValidator(pubkey, v.VotingPower, "")
+			vUpdates = append(vUpdates, update)
+			app.log.Info("Updated voting power of validator", logging.String("tmKey", v.TmPubKey), logging.Int64("votingPower", v.VotingPower))
+		}
+
+		resp = tmtypes.ResponseEndBlock{
+			ValidatorUpdates: vUpdates,
+		}
+	}
+	return ctx, resp
 }
 
 // OnBeginBlock updates the internal lastBlockTime value with each new block.